Bug fixes and performance improvements

+++ locallib.php
@@ -476,7 +476,7 @@ function seconds_to_time($t,$f=':'){
 	if($t < 0){
 		return "00:00:00";
 	}
-	return sprintf("%02d%s%02d%s%02d", floor($t/3600), $f, ($t/60)%60, $f, $t%60);
+	return sprintf("%02d%s%02d%s%02d", (int) floor($t/3600), $f, (int) ($t/60)%60, $f, (int) $t%60);
 }
